refactor(local-mcp): simplify classification and self-fetch imported servers
Cleanups from a /simplify pass:

- Move the McpServerConfig -> LocalMcpTransport normalization into the
  agent's mcp-config.ts beside the parser; the public export now
  returns the stable @posthog/shared descriptor type instead of
  leaking the Claude SDK config shape, and the workspace-server
  service becomes a one-line delegate. Drops the duplicate
  ClaudeJsonMcpScope type in favor of LocalMcpServerScope.
- Slim LocalMcpCloudClassification to what consumers read: drop the
  derivable transportType and the never-rendered scope; collapse the
  two identical invalid_url branches into a parseHttpUrl helper.
- useTaskCreation now self-fetches importable servers (matching the
  additionalDirectories pattern) instead of TaskInput threading them
  through props — ChannelHomeComposer cloud tasks now get them too.
  The hook moves to features/local-mcp/, returns a stable empty-array
  identity, and loses its unreachable service guard.

Generated-By: PostHog Code
Task-Id: 7c87c3a4-0be6-475e-a8ec-269140ded301
