+++ packages/ui/src/features/task-detail/hooks/useTaskCreation.ts
@@ -13,7 +13,6 @@ import { useHostTRPC, useHostTRPCClient } from "@posthog/host-router/react";
 import {
   type Adapter,
   ANALYTICS_EVENTS,
-  type CloudMcpServerImport,
   PROJECT_BLUEBIRD_FLAG,
   type TaskCreationInput,
   type WorkspaceMode,
@@ -35,6 +34,7 @@ import { titleAttachmentStoreApi } from "../../../shell/titleAttachmentStore";
 import { useAuthStateValue } from "../../auth/store";
 import { assertCloudUsageAvailable } from "../../billing/preflightCloudUsage";
 import { useUsageLimitStore } from "../../billing/usageLimitStore";
+import { useLocalMcpCloudServers } from "../../local-mcp/useLocalMcpCloudServers";
 import {
   contentToPlainText,
   contentToXml,
@@ -88,8 +88,6 @@ interface UseTaskCreationOptions {
    * whether it needs one and attaches it lazily.
    */
   allowNoRepo?: boolean;
-  /** Importable local MCP servers to forward into a cloud run's sandbox. */
-  importedMcpServers?: CloudMcpServerImport[];
   onTaskCreated?: (task: Task) => void;
   /**
    * Side effect run with the created task in addition to (not instead of)
@@ -179,7 +177,6 @@ export function useTaskCreation({
   channelName,
   channelId,
   allowNoRepo,
-  importedMcpServers,
   onTaskCreated,
   onTaskCreatedEffect,
 }: UseTaskCreationOptions): UseTaskCreationReturn {
@@ -196,6 +193,9 @@ export function useTaskCreation({
     useState<string[] | null>(null);
   const additionalDirectories =
     additionalDirectoriesOverride ?? defaultAdditionalDirectories;
+  // Importable local MCP servers for cloud runs, self-fetched like the
+  // additional-directory defaults above rather than threaded in by callers.
+  const localMcpServers = useLocalMcpCloudServers(workspaceMode === "cloud");
   const taskService = useService<TaskService>(TASK_SERVICE);
   const clearTaskInputReportAssociation = useTaskInputPrefillStore(
     (s) => s.clearReportAssociation,
@@ -357,7 +357,9 @@ export function useTaskCreation({
           autoPublishCloudRuns: settings.autoPublishCloudRuns,
           rtkEnabledCloud: settings.rtkEnabledCloud,
           allowNoRepo,
-          importedMcpServers,
+          importedMcpServers: localMcpServers.flatMap((server) =>
+            server.remote ? [server.remote] : [],
+          ),
         });
 
         if (executionMode) {
@@ -528,7 +530,7 @@ export function useTaskCreation({
       allowNoRepo,
       bluebirdEnabled,
       personalChannel?.id,
-      importedMcpServers,
+      localMcpServers,
       clearTaskInputReportAssociation,
       invalidateTasks,
       onTaskCreated,
